fix bound color

+++ server.py
@@ -4,20 +4,8 @@ from grid import rendering
 async def handle_echo(reader, writer):
     data = await reader.read(100)
     message = data.decode()
-
     command = message.split(',')
-
-    rendering(command[0], int(command[1]), int(command[2]))
-
-    # addr = writer.get_extra_info('peername')
-
-    # print(f"Received {message!r} from {addr!r}")
-    # print(f"Send: {message!r}")
-    # writer.write(data)
-    # await writer.drain()
-
-    # print("Close the connection")
-    # writer.close()
+    rendering(command[0], command[1], float(command[2]) )
 
 async def main():
     server = await asyncio.start_server(
